First step of refactor intended to make code more readable/maintainable.
Creates an explicit dimension type with the intent to add more dimensioning options (both fixed length and generator) in the future.
Also refactors the grid functionality for a cleaner separation between grid declaration (lazy) and element materialization, with an eye towards supporting sampling and transformation into other frameworks in the future.